Adjust to sort lists with different values

+++ Lista1/src/MergeSort.java
@@ -2,10 +2,10 @@ import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
 
-public class MergeSort<Integer> {
-    public ArrayList<Integer> mergeSort(List<Integer> list, int startIndex, int endIndex){
+public class MergeSort<Double> {
+    public ArrayList<Double> mergeSort(List<Double> list, int startIndex, int endIndex){
         if (startIndex == endIndex) {
-            ArrayList<Integer> result = new ArrayList<>();
+            ArrayList<Double> result = new ArrayList<>();
             result.add(list.get(startIndex));
             return result;
         }
@@ -13,18 +13,18 @@ public class MergeSort<Integer> {
         return merge(mergeSort(list, startIndex, splitIndex), mergeSort(list, splitIndex + 1, endIndex));
     }
 
-    private ArrayList<Integer> merge(List<Integer> left, List<Integer> right) {
-        ArrayList<Integer> result=new ArrayList<>();
-        Iterator<Integer> l = left.iterator();
-        Iterator<Integer> r = right.iterator();
-        Integer elemL=null;
-        Integer elemR=null;
+    private ArrayList<Double> merge(List<Double> left, List<Double> right) {
+        ArrayList<Double> result=new ArrayList<>();
+        Iterator<Double> l = left.iterator();
+        Iterator<Double> r = right.iterator();
+        Double elemL=null;
+        Double elemR=null;
         boolean contL;
         boolean contR;
         if(contL=l.hasNext()) elemL=l.next();
         if(contR=r.hasNext()) elemR=r.next();
         while (contL && contR) {
-            if ((int)elemL<=(int)elemR){
+            if ((double)elemL<=(double)elemR){
                 result.add(elemL);
                 if(contL=l.hasNext()) elemL=l.next();
                 else result.add(elemR);
